Added a separate component for the Tabs
